Cuda : Statistic Qualifier

+++ src/components/cuda/cupti_events.h
@@ -19,11 +19,11 @@ int cuptie_init(void);
 int cuptie_shutdown(void);
 
 /* native event interfaces */
-int cuptie_evt_enum(uint64_t *event_code, int modifier); 
-int cuptie_evt_code_to_descr(uint64_t event_code, char *descr, int len);
-int cuptie_evt_name_to_code(const char *name, uint64_t *event_code);
-int cuptie_evt_code_to_name(uint64_t event_code, char *name, int len);
-int cuptie_evt_code_to_info(uint64_t event_code, PAPI_event_info_t *info);
+int cuptie_evt_enum(uint32_t *event_code, int modifier); 
+int cuptie_evt_code_to_descr(uint32_t event_code, char *descr, int len);
+int cuptie_evt_name_to_code(const char *name, uint32_t *event_code);
+int cuptie_evt_code_to_name(uint32_t event_code, char *name, int len);
+int cuptie_evt_code_to_info(uint32_t event_code, PAPI_event_info_t *info);
 
 /* profiling context handling interfaces */
 int cuptie_ctx_create(void *thr_info, cuptie_control_t *pctl);
